Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/1364717

+++ src/trace_processor/db/column.cc
@@ -186,7 +186,7 @@ void Column::FilterIntoNumericSlow(FilterOp op,
         // for this function even though the LHS of the comparator should
         // actually be |v|. This saves us having a duplicate implementation of
         // the comparision function.
-        return -compare::LongToDouble(long_value, v);
+        return -compare::LongToDouble(long_value, static_cast<double>(v));
       };
       FilterIntoNumericWithComparatorSlow<T, is_nullable>(op, rm, fn);
     } else {
